website_app/requirements.txt
Flask
pymongo
azure-storage-blob
gunicorn
